backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.35.71
botocore==1.35.71
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
coverage==7.6.8
cryptography==44.0.0
dataclasses-json==0.6.7
defusedxml==0.8.0rc2
Deprecated==1.2.18
distro==1.9.0
Django==5.1.3
django-appconf==1.1.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-ipware==7.0.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-recaptcha==0.2.0
djangorestframework-simplejwt==5.3.1
docx2txt==0.9
drf-spectacular==0.28.0
durationpy==0.9
exceptiongroup==1.2.2
Faker==33.1.0
fastapi==0.115.9
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
itsdangerous==2.0.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.4.2
kubernetes==32.0.1
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pgvector==0.3.6
pillow==11.0.0
posthog==3.25.0
prompt_toolkit==3.0.48
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.7
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-ipware==3.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.10.4
sentry-sdk==2.26.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.7
SQLAlchemy==2.0.36
sqlparse==0.5.2
starlette==0.45.3
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.9.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
